+++ selene-lib/src/rules.rs
@@ -18,6 +18,7 @@ pub mod parenthese_conditions;
 pub mod shadowing;
 pub mod standard_library;
 pub mod suspicious_reverse_loop;
+pub mod type_check_inside_call;
 pub mod unbalanced_assignments;
 pub mod undefined_variable;
 pub mod unscoped_variables;