+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 ### Added
+- Added `type_check_inside_call` lint for checking `type(foo == "type")` instead of `type(foo) == "type"`.
 - Added `NumberRange` to the Roblox standard library.
 - Added `string.split` to the Roblox standard library.
 - Added `table.find` to the Roblox standard library.